asdf

+++ string.cpp
@@ -79,6 +79,6 @@ bool FGETS(String *line, uint LINE_MAX_LENGTH, FILE *file) {
     ASSERT(line);
     ASSERT(file);
     bool result = fgets(line->data, LINE_MAX_LENGTH, file);
-    line->length = strlen(line->data);
+    line->length = uint(strlen(line->data));
     return result;
 }
